Optimizing

+++ source/UDP.hpp
@@ -91,7 +91,7 @@ public:
         }
         return true;
     }
-    int receive(void* const buf_ptr, const size_t buf_size) {
+    ssize_t receive(void* const buf_ptr, const size_t buf_size) {
         auto output = recv(sock, buf_ptr, buf_size, 0);
         if (output == -1) {
             std::cout << "receive error" << std::endl;
